requirements.txt
Flask==2.0.2
Flask_Bootstrap==3.3.7.1
psycopg2_binary==2.9.2
python-dotenv==1.0.0
requests==2.21.0
SQLAlchemy==1.4.27
tenacity==8.0.1
